Changed int to doubles in ranking-package. Added an empty test-class for Scores

+++ searchengine/src/ranking/IDFScore.java
@@ -6,7 +6,7 @@ import index.Index;
 public class IDFScore implements Score {
 
     @Override
-    public int getScore(String word, Website site, Index index) {
+    public double getScore(String word, Website site, Index index) {
         return 0;
     }
 }
