+++ searchengine/src/ranking/TDIDFScore.java
@@ -6,7 +6,7 @@ import index.Index;
 public class TDIDFScore implements Score {
 
     @Override
-    public int getScore(String word, Website site, Index index) {
+    public double getScore(String word, Website site, Index index) {
         return 0;
     }
 }
